Create MMScriptStart.ps1
This script will first check if any objects already in maintenance mode in SCOM or not for a server. If any object in
maintenance mode, this script will fetch the information from SCOM and kept it in AlreadyMM.txt file And 
then it will put entire server in maintenance mode.